added concurrency update with optimistic locking

+++ api/src/main/java/com/iwaodev/application/service/CategoryServiceImpl.java
@@ -14,9 +14,11 @@ import com.iwaodev.infrastructure.model.Category;
 import com.iwaodev.ui.criteria.category.CategoryCriteria;
 import com.iwaodev.ui.criteria.category.CategoryQueryStringCriteria;
 
+import com.iwaodev.util.Util;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.OptimisticLockingFailureException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -24,21 +26,23 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 @Service
 @Transactional
 public class CategoryServiceImpl implements CategoryService {
 
   private static final Logger logger = LoggerFactory.getLogger(CategoryServiceImpl.class);
 
+  @Autowired
   private CategoryRepository repository;
 
+  @Autowired
   private CategorySpecificationFactory specificationFactory;
 
   @Autowired
-  public CategoryServiceImpl(CategoryRepository repository, CategorySpecificationFactory specificationFactory) throws Exception {
-    this.repository = repository;
-    this.specificationFactory = specificationFactory;
-  }
+  private HttpServletRequest httpServletRequest;
 
   public Page<CategoryDTO> getAll(CategoryQueryStringCriteria criteria, Integer page, Integer limit,
       CategorySortEnum sort) throws Exception {
@@ -98,7 +102,16 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     Category category = targetEntityOption.get();
-    
+
+    // version check for concurrency update
+    String receivedVersion = this.httpServletRequest.getHeader("If-Match");
+    if (receivedVersion == null || receivedVersion.isEmpty()) {
+      throw new AppException(HttpStatus.BAD_REQUEST, "you are missing version (If-Match) header.");
+    }
+    if (!Util.checkETagVersion(category.getVersion(), receivedVersion)) {
+      throw new AppException(HttpStatus.PRECONDITION_FAILED, "the data was updated by others. please refresh.");
+    };
+
     // duplication check
     if (this.repository.isOthersHaveName(category.getCategoryId(), criteria.getCategoryName())) {
       throw new AppException(HttpStatus.BAD_REQUEST, "the category name already taken.");
@@ -114,13 +127,22 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     // map criteria to entity
+    // issue-k0rdIGEQ91U
     Category updateEntity = CategoryMapper.INSTANCE.toCategoryEntityFromCategoryCriteria(criteria);
+    category.update(updateEntity);
 
-    // save it
-    this.repository.save(updateEntity);
+    Category updatedCategory;
+    try {
+      // save it
+      // issue-j845jixIPCn
+      // don't forget flush otherwise version number is updated.
+      updatedCategory = this.repository.saveAndFlush(category);
+    } catch (OptimisticLockingFailureException ex) {
+      throw new AppException(HttpStatus.CONFLICT, "the data was updated by others. please refresh.");
+    }
 
     // map entity to dto and return it.
-    return CategoryMapper.INSTANCE.toCategoryDTO(updateEntity);
+    return CategoryMapper.INSTANCE.toCategoryDTO(updatedCategory);
   }
 
   @Override
@@ -132,10 +154,23 @@ public class CategoryServiceImpl implements CategoryService {
     if (targetEntityOption.isPresent()) {
       Category targetEntity = targetEntityOption.get();
 
+      // version check for concurrency update
+      String receivedVersion = this.httpServletRequest.getHeader("If-Match");
+      if (receivedVersion == null || receivedVersion.isEmpty()) {
+        throw new AppException(HttpStatus.BAD_REQUEST, "you are missing version (If-Match) header.");
+      }
+      if (!Util.checkETagVersion(targetEntity.getVersion(), receivedVersion)) {
+        throw new AppException(HttpStatus.PRECONDITION_FAILED, "the data was updated by others. please refresh.");
+      };
+
       if (targetEntity.getProducts().size() > 0) {
         throw new AppException(HttpStatus.BAD_REQUEST, "cannot delete this category since it has products.");
       }
-      this.repository.delete(targetEntity);
+      try {
+        this.repository.delete(targetEntity);
+      } catch (OptimisticLockingFailureException ex) {
+        throw new AppException(HttpStatus.CONFLICT, "the data was updated by others. please refresh.");
+      }
     }
   }
 }
